Album-Search app start
Album-Search uses search form returns album data in JSON format based
on the form submission. Using the API data, dynamically created HTML
displays an album's title, artist, and album art image.

—————————————————

Sample HTML, CSS, and images committed to repository.

The once empty JS file has some pseudocode written for sending an AJAX
request to the Spotify API.

Spotify user is logged in so that use of Spotify API is granted.

JS is written to access the value of the form on submit while
preventing the default action of reloading the page.

$.getJSON() method’s syntax is beginning to be structured.